Let Travis-ci run Google App Engine tests
Travis build downloads the GAE sdk and can now run the GAE test. To
allow the same tests to run on python2.6 and 2.7, this commit include a
test runner script similare to nosegae except it doesn't break on
python 2.6.
Nosetests should still be used for development since it collects logs
and display them on error.